WOM-115 1d array timestamp-value tuple and simple reader
Updates the logger class to store data as SourceTimestamp Value tuples
in a 1 dimensional array with a dataset per node with a stub for node
value type retrieval.
Also adds a simple reader class to plot data from a dataset (node) in a
HDF5 file, allowing the user to choose a datetime range to plot, for use
with a Jupyter notebook. N.B only tested with nodes that have value type
double.