WOM-152 Connect datalog to sculib types and enums
Removes the stub functions for getting the type and the enum strings
from the datalogger and replaces them with sculib calls.

Also fixes the sculib types function as multiple calls to the
load_data_type_defintions() in the connect method in the same
interpreter were giving different results, causing errors in testing.