WOM-236: GUI controller capture and handle RuntimeError when connecting to server
The instantiation of sculib creates the connection. All asyncua exceptions inherit from RuntimeError so this will catch any form of connection error.
If sculib raises an exception during construction of the scu object, then we delete the object because we can't be sure of its state and re-raise the exception