WOM-339: Add tests of opcua commands through GUI
The tests is run in the CI with the model.run_opcua_command method
mocked. On a local machine the tests will try to connect to a running
CETC simulator.

Further work to be done - only the basic happy paths are tested, but
the ground work is in place to easily expand the tests with sad paths
and edge cases.