WOM-: Refactor SCU into a base and child class
AsyncOPCUAClient is a base class that is usable on its own as a basic
client with an event loop and methods to handle a connection with user
authentication.

SteeringControlUnit inherits from the new base and extends it with all
the functionality specific to the Dish Structure Controller's PLC and
CETC sim.