WOM-402: Remove Root tree scan and its resulting dicts
Scanning the entire Root tree has no use case at the moment, and is just
duplicating the other existing node dicts with a lot of other mostly
useless nodes.

The generate_node_dicts_from_server() method can be used to scan a
specific node's tree and get dicts of it, if needed by an end-user.

Also removed the parameters_commands dict, as it is empty anyway.