WOM-471: Swop sculib with ska-mid-dish-steering-control package
The disq package now imports the SCU class, generator and constants from
ska-mid-dish-steering-control and exposes them as if they are part of
disq - thus not breaking the existing interface.

Removed all the constants (enums) already in
ska-mid-dish-steering-control from disq.constants. Also copied the
sculib module docstring to top of sculib.rst, as it's now unavailable to
import directly into the docs.