WOM-427: Ignore event updates with 'None' value to prevent TypeErrors
If the PLC is reset while the GUI is connected to it, in the short time
it takes to reset, some event updates is received with a value of
'None', which caused the GUI to crash when the 'None' is casted to an
int in the widget update functions.

Ignoring the 'None' values prevents the TypeErrors. A warning is logged
when 'None' is received.