WOM-401: Change individual axis position & velocity widgets to spinboxes
Clicking on the up/down buttons of the position spinboxes immediately
sends a slew command for quick adjustment. The position and velocity
inputs now have allowed  ranges according to the limits specified in the
ICD. In other words, input validation is now hard-coded into the
widgets, instead of being checked in the slot function.